+++ apps/aptos/components/Pools/components/PoolCard/CakeCollectModal.tsx
@@ -2,44 +2,23 @@ import { useQueryClient } from '@pancakeswap/awgmi'
 import { Pool } from '@pancakeswap/uikit'
 import useHarvestFarm from 'components/Farms/hooks/useHarvestFarm'
 import { CAKE_PID } from 'components/Pools/constants'
+import cakePoolRelatedQueries from 'components/Pools/utils/cakePoolRelatedQueries'
 import useActiveWeb3React from 'hooks/useActiveWeb3React'
 import { useCallback } from 'react'
-import { FARMS_NAME_TAG, FARMS_USER_INFO } from 'state/farms/constants'
 import CollectModalContainer from './CollectModalContainer'
 
-const CakeCollectModal = ({
-  earningTokenAddress,
-  ...rest
-}: React.PropsWithChildren<
-  Pool.CollectModalProps & {
-    earningTokenAddress: string
-    stakingTokenAddress: string
-  }
->) => {
+const CakeCollectModal = ({ earningTokenAddress = '', ...rest }: React.PropsWithChildren<Pool.CollectModalProps>) => {
   const { onReward } = useHarvestFarm(CAKE_PID, earningTokenAddress)
   const queryClient = useQueryClient()
-  const { account, networkName } = useActiveWeb3React()
+  const { account } = useActiveWeb3React()
 
   const onDone = useCallback(() => {
     queryClient.invalidateQueries({
-      predicate: (query) => {
-        const queryObject = query.queryKey[0]
-
-        const isMasterchefQuery =
-          queryObject?.entity === 'accountResource' && queryObject?.resourceType === FARMS_NAME_TAG
-
-        const isPoolUserInfoHandleQuery =
-          queryObject?.entity === 'tableItem' && queryObject?.data?.valueType === FARMS_USER_INFO
-
-        return isMasterchefQuery || isPoolUserInfoHandleQuery
-      },
+      predicate: cakePoolRelatedQueries(account),
     })
-    queryClient.invalidateQueries({
-      queryKey: [{ entity: 'accountResources', networkName, address: account }],
-    })
-  }, [account, networkName, queryClient])
+  }, [account, queryClient])
 
-  return <CollectModalContainer onDone={onDone} onReward={onReward} {...rest} />
+  return <CollectModalContainer {...rest} onDone={onDone} onReward={onReward} />
 }
 
 export default CakeCollectModal
