+++ apps/aptos/components/Pools/utils/getSecondsLeftFromNow.ts
@@ -1,0 +1,7 @@
+const getSecondsLeftFromNow = (timestamp: number) => {
+  const now = Math.floor(Date.now() / 1000)
+
+  return Number.isFinite(timestamp) && timestamp < now ? now - timestamp : 0
+}
+
+export default getSecondsLeftFromNow
